Lots of changes - real revisionnotes will follow ...
- "Meaningless" urls for services. Ie. the mapping from endpoint to entityID, binding and service
is done by a map. The map is created from metadata. Dutch scoping can thus be accomplished by making
proxy entities in metadata for each real entity.

- Multi federation support. Corto now supports multible federations per installation. When importing
metadata the federation it belongs can be stated.

- Import of std. metadata. Almost all configuration of corto is now done by importing std. metadata.
Corto specific configuration is done using SAML V2.0 Metadata Extension for Entity Attributes.

- Proxying can now be done with different entities on each side. Ie. the idp side of the proxy can
be configured to use a specific (hosted) sp.

- JSON bindings always use openssl compatible aes-256-cbc encryption. A passphrase is configured and
is it used to generate both a key and an initialisation vector.